refactor(nealsdata): scaffold converter package + Excel loader
